Update patch set 1
Patch Set 1:

I *think* this is causing vLB to fail. When I compare against the wiki CSIT from Amsterdam it seems that this is the only missing field. We are trying to track down via SO if that is the reason why. Their logs seems to not show reasons why.

Patch-set: 1