Update patch set 2
Patch Set 2:

Pam, Liam, first is 1) common, 2) drools-pdp, 3) drools-applications, 4) engine.   If you remember we had a cyclical dependency in the past with drools-pdp needing drools-applications, and drools-applications needing drools-pdp libraries we used.   

This was triggered because of removing the build portion from the docker image.   To resolve this cyclical dependency I had to make drools-pdp agnostic of prepackaging drools-applications deliverables.   

The solution was to put hooks into external scripts policy_vm_init that then I replaced with the drools-preinstall hook to drive the download of the matching drools-applications zip at installation time.   It also has the good side-effect that the drools-pdp can be used standalone with no applications pre-configured if desired for their own means.

Patch-set: 2