Update patch set 1
Patch Set 1:

> I thought the issue was also that VFC shouldn't require the
 > property. When I was working with Yang we were modifying the
 > properties file to bypass some issues when we uncovered this
 > exception.

So here's the question: if we make that change, do we still need to catch the exception?  If not, then I'll abandon this review.  On the other hand, if we still want to catch the exception, then we should go ahead with this review.  In either case, I planned to make the other changes (i.e., don't require the property) via another review.

Patch-set: 1