Update patch set 1
Patch Set 1:

Ive updated the poms. 
There will still be one compile problem:
PolicyControllerFeatureAPI in drools-pdp is implemented in ControlLoopMetricsFeature in drools-applications, the signature of the beforeDeliver(...) method has changed due to the move of ...endpoints.event.comm.Topic.CommInfrastructure from drools to common. So there will be a mismatch between the interface and implementing class until drools-pdp is built

Patch-set: 1
Label: Code-Review=0