Update patch set 2
Patch Set 2:

If the need arises to change to use a database for all users then the existing implementation class can be updated without impacting on the users. I would however consider that it is not necessarily a bad thing to impact all uses for such a change, so as each use can consider if the changed implementation is appropriate for that use. 
Aside from that, and I accept this is subjective, I think an interface should be just an interface, i.e an abstraction. Adding implementation code into the interface dilutes the concept of it being an abstraction.

If non singleton behaviour is required for a particular use, then a separate implementation of the interface can be provided that does not have the singleton behaviour, with the users free to pick the implementation that suits their needs.
In a scenario where somebody could instantiate a factory directly using new, then the static final factory member will contain a reference to the wrong factory for the classes instantiated by that factory, leaving the door open for errors

Regarding the "Indexed" names proliferating in user code, perhaps we could come up with a name that would read better, maybe CachingXXXFactory or SingletonCachingXXXFactory or...

Patch-set: 2