Update patch set 2
Patch Set 2:

Michael, the fact that somebody may or may not like final static data members in an interface does not justify design changes.   A simple change of the “interface” to an “abstract class” would do it in that case.    I have an issue with the scope of changes as the original design is altered and constrained.   Again, to summary my thoughts in a list:

1.	Issue 1:  A factory implementation is now a singleton.   With the current changes, factory implementations are singletons, they have a "getInstance()" and there can only be used in a singleton way.   This is a change from the original implementation and it has constrained the original design intent, as the for example IndexedUebTopicSinkFactory and all those factories could be instantiated as many as you want with “new”, and it may make sense to have them as such, when the PDP-D supports same topics, in different  sever clusters, ie. topic X in host 1, and topic X in host 2.   Right now, only one instance on per application space is accessible.   This is a change and a constraint in the design. 
2.	Issue 2:  Users of the code were coding to an interface, now the implementation leaks into their code.
3.	Issue 3:  Several hundreds of unnecessary lines of code inspections have been added.   Only thing necessary would have been to change the package name in the import, and change the pom.xml to realize the moving of the code change.     This work goes far beyond that, and impacts additional hundreds of lines of codes on the users of these classes.
4.	Issue 3:  Readability by users of class decreases, as the original intent that the implementation details get hidden, now the “Indexed” .. etc .. has to carry forward in all the uses.   Further, all name changes, have to be carried onwards by any other project outside policy that uses it, etc ..

I am sorry I didn’t catch this early, but was just thinking this was a plain routine move of code.   

I understand  that we want to merge the code in the base, and you’re going on vacations.  Not sure, what’s the best way to move forward, but I’m open to solutions to merge this with some contingencies on some work being done on as a follow up on this regard as a consequence of my comments.

Patch-set: 2