Update patch set 10
Patch Set 10:

We traditionally were using in drools-pdp and drools-applications for scripts, scripts with no extensions, so their invocation could be kept generic regardless the scripting language used underneath.   The other convention is to use "-" instead of "_".   I know that the previous .jmx script from amsterdam leaked with ".sh", but better to stick to common conv.  Is there a reason to deviate from the original convention with this script?

Patch-set: 10