Update patch set 10
Patch Set 10: -Code-Review

> We traditionally were using in drools-pdp and drools-applications
 > for scripts, scripts with no extensions, so their invocation could
 > be kept generic regardless the scripting language used underneath. 
 >  The other convention is to use "-" instead of "_".   I know that
 > the previous .jmx script from amsterdam leaked with ".sh", but
 > better to stick to common conv.  Is there a reason to deviate from
 > the original convention with this script?

It follows the same conventions as testsuites/stability.  However, I can remove the ".sh" extension, if desired.  Regarding "-" instead of "_", though, I find "_" much easier to use, as the clipboard "selection" operation in putty, double-click, does not span hyphens, while it does span underscores.  If the convention is to use "-", then that particular convention should be changed going forward - it's a major pain to not be able to use double-click.  (IMO, this is a handicap accessibility issue.)

Patch-set: 10