Update patch set 10
Patch Set 10: Code-Review-1

Let's stick with the original convention for these repos to try to keep a consistent apperance for files of the same type.  We can have a separate general conversation to change it for all files, but would be an all-or-nothing changing at the same time.  Key thing is consistency.  As a note, I don't experience the double-click selection behavior that you're mentioning.

Patch-set: 10
Label: Code-Review=-1