Update patch set 3
Patch Set 3:

Hmm - not so simple to fix guard tests. I will finish the checkstyle and sonar stuff first. Then go back and address the code - which isn't quite right. Its throwing snakeyaml exceptions, which it shouldn't. The tests are correct - there should not be a fail but rather a check to see if the right exception was thrown.

Patch-set: 3