fix new sonar issues with prov-status check
Changes made while adding the prov-status check triggered some sonar
thresholds in ControlLoopEventManager.  Updated the code to resolve
them.  One of the changes that had been introduced was to check for
the existence of both the is-closed-loop-disabled AND the prov-status
fields.  However, if one is there then the other should be there, thus
the second check is redundant and was removed, which should reduce the
cyclomatic complexity back under the threshold.

Change-Id: If0ce74cedbf947c32cae3df3374cda6ba1305cd2
Issue-ID: POLICY-964
Signed-off-by: Jim Hahn <jrh3@att.com>