move AAI named query
The AAI vserver named-query was being done as a preliminary step for the
SO request.  However, it will now be needed during the guard query, which
happens much earlier.  Thus it has been moved from the SO classes into
the event manager class, where it is retrieved in a lazy fashion and
cached.
Also removed some trailing white space.
Update licenses.
Make test method private.

Change-Id: Id7353a2e0f2d42601d5142cb92961528fa21aa94
Issue-ID: POLICY-1011
Signed-off-by: Jim Hahn <jrh3@att.com>