use new vf-walk code in SO request
Methods were added to AaiNqResponseWrapper to extract VF modules
and generate the new VF module name.  This step modifies the SO
code to use those methods.  It was also determined that the SO
code needs the VF module container object, not just the VF module,
itself.  As a result, AaiNqResponseWrapper was modified to return
a list of the containers instead of a list of the VF modules.

Also modified the AAI simulator to return two VF modules in the
response to the vserver named query.  As part of that, a method
was added to the simulator so that JSON responses can be read from
files rather than having to convert them to Java Strings.

Modified simulator to work even if vnf-id is null.

Change-Id: I68fdf07ea80ee0daf9e16403e35b11710315a8a8
Issue-ID: POLICY-1037
Signed-off-by: Jim Hahn <jrh3@att.com>