Update patch set 2
Patch Set 2:

(6 comments)

> Piyush- you will probably have to rebase this as we just pulled in
 > a big cleanup of the checkstyle code. Sorry, blame me :)

No problem. I have already rebased it and committed the changes for some of the review comments.
But it looks like in the latest code there are compilation issues in the simulators project due to missing org.onap.policy.drools.utils.logging.LoggerUtil and causing the build failure.

Patch-set: 2