Update patch set 6
Patch Set 6:

> (1 comment)
 > 
 > I'm ok as is. I'm not sure what your comment about A&AI simulator
 > means Jim.

As this is new code, I'd rather not have to go back and rename things later, thus would prefer the fields to be renamed to result & message (or reuse an existing Pair - why reinvent the class and then have to add junit tests for it? though, hopefully, it's already covered by the junit tests for the containing class).
As for A&AI, I had requested earlier that all of the A&AI stuff be removed, as it isn't needed.  The response was to leave it in, in case it's needed later.  I'm OK with that, but in that case, ALL of the A&AI stuff should be left in, otherwise someone will have to track down why it doesn't work later, if it is put back into use.  Hence, if we're going to leave it building the A&AI simulator, then we should also leave the A&AI environment properties, which are needed to connect to the simulator.

Patch-set: 6