Update patch set 1
Patch Set 1:

(7 comments)

Joshua I was anticipating that a whole new separate template would be created for this to support the embedded PDP until we fixed the guard implementation.

Nevertheless, let's review it and see. I would rather not have to add a modification to the YAML just to specify coordination true or false.

Let's also be very clear that the embedded XACML PDP is POC only, and should be reviewed to be removed in a subsequent release.

Patch-set: 1