Update patch set 1
Abandoned

Thanks to everyone for their reviews thus far.  As mentioned in my code review, I'm going to move the demo code from template.demo to a new template.clc and thereby minimize the changes to the rest of the code base (yaml spec, embedded PDP, etc.).  I will mark this as abandoned and submit a new Gerrit review in a couple of hours.

Patch-set: 1
Status: abandoned