fix NPE from rule timers
This problem occurs in rules that have "timer(expr)" expressions before
the "when" clause.  If a working memory element, other than the timer
object, appears in the "when" clause and is updated BY ANY RULE, then
an NPE is thrown.  The solution is to segregate the objects so that
rules having "timer(expr)" expressions only refer to one object, the
XxxTimer object.  In particular, the rule simply sets an "expired" flag
in the XxxTimer object and then all other rules check that flag, rather
than having their own "timer(expr)" expressions.

Change-Id: I9e6c20ee46af99234daee8ece4862edbd41ba714
Issue-ID: POLICY-1106
Signed-off-by: Jim Hahn <jrh3@att.com>