Fix sonar critical
Sonar complains that the password is an actual password when
in fact we are doing exactly the fix they suggest.

I seem to not have access to resolve it as a false positive.
So it seems the better way to do this is rename this to
reflect that it is actually a property value.

More checkstyle nightmares as the clearly defined suppression
filter for ONLY abbreviations and type names now also ignores
120 character line problems.

Issue-ID: POLICY-1135
Change-Id: I2e2a36ed8d7f5c95e912df4f6ed74983cee8a3dd
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>