Changed actor from "MSO" to "SO"
Also added "VF Module Delete" to the list of recipes for "SO" accepted
by the compiler.
Removed "AOTS" and "SDNO" from the list of accepted actors, as code does
not exist for those actors.
Update remaining actors and recipes.  Left "AAI" in the list for now,
though it has no recipes listed (and no actor.aai code).
Also removed "AAI" from the list of accepted actors.

Change-Id: I9db56715d8bea3cf86d2112df1e0086777863192
Issue-ID: POLICY-1239
Signed-off-by: Jim Hahn <jrh3@att.com>