Update patch set 2
Patch Set 2:

Thank you Jim and Josh for looking in detail into this code review.

Given at this stage of the release where we are, it's critical to make sure that there is no breakage associated with these changes in the runtime operation of the use cases.   This is unless there is an extension on the release dates (Pam can weigh in).  

If we are not confident about not incurring in additional breakage, and there is no dates extension, it would be better to complete the release with an exception mentioning a "restart" workaround for the upgrade case that Jim is trying to handle.

Feedback?

Patch-set: 2