Update patch set 2
Patch Set 2: -Code-Review

What I can say is that Jim's code looks good to me based on the review that I've done and the exchange with Jim (in fact, having caught only two minor bugs gives me a bit extra faith in the code). That said, Jim will need to weigh in on his level of confidence on this not incurring in additional breakage.

Patch-set: 2