Update patch set 2
Patch Set 2: Code-Review-2

I am really uncomfortable with having to restart the Drools controller manually everytime a rule is deleted/updated. But Jorge has a point. Please let me work with Integration team to determine if we can push this in now and get it tested.

I will undo the -2 when we agree, please keep the review open.

Patch-set: 2