Update patch set 2
Patch Set 2:

The engine does not get corrupted for what I see, able to process messages.
The only issue is that the in the delete control loop.  Those control loops Param facts don't get retracted.
In any case, I think it is an improvement.   The delete control loop in theory is a low runner scenario, in comparison with the update.

Patch-set: 2