Update patch set 1
Patch Set 1:

> Build Failed
 > 
 > https://jenkins.onap.org/job/policy-drools-applications-master-verify-java/874/
 > : FAILURE
 > 
 > maven-compiler-plugin testCompile goal has failed to compile tests.
 > Please review the compilation issues in the log ( https://jenkins.onap.org/job/policy-drools-applications-master-verify-java/874/
 > )
 > 
 > Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/policy-drools-applications-master-verify-java/874

The build failed due cross project dependency. I modified the code in drools-pdp in prev commit., and these objects are used in test cases of drools-applications, and hence the test class failed to compile. 
i did not take care of this cross compiling. actually this is first time i encountered the case, where a class defined in drools-application is used outside of this project in another project such as drools-pdp. 
Is this type of dependency valid ? in which case, if we change a class in one project, we might have to compile all the other projects.

Patch-set: 1
Reviewer: Gerrit User 2543 <2543@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0