Update patch set 5
Patch Set 5:

Jim, what's the reason for the swapping of arguments Assert.assertEquals()? 

It makes sense with the String.equals comparison, as you want to preceded it with the constant to prevent passing null pointers, but this is not that case, and it is instead a method call into the Junit Assert class, rather than String comparison, and the same concern does not apply.   If you look into the Junit library, it does a Strings.valueOf on each argument and then it compares, so will never run into that kind of situation.

I guess, I would like to understand what's the rule or rationale dictating argument order into arbitrary method calls into 3rd party libraries, as it is an arbitrary method call to a 3rd party method that takes 2 strings, and not a String.equals comparison.

Patch-set: 5