Update patch set 1
Patch Set 1:

> Ok this was put it so the latest Jackson could be used and remove
 > older version being used. (BTW I will be updating the XACML PDP to
 > remove jackson also). Is this needed to be put back in because JVM
 > now cannot find jackson for XACML to use?

Not sure I fully understand the comment.  From the mvn dependency tree, it does not appear that "xacml-pdp" uses jackson directly, while 
"xacml" does reference jackson.  This runs in a standalone install, so didn't think jackson was really needed, particularly since the code does not serialize/deserialize any of the xacml data structures.  Nevertheless, may it's happy because it's still getting jackson due to policy-endpoints or policy-management.
So, I can leave this change in for now, and fix things later once jackson is removed from policy-endpoints and policy-management, or I can abandon this change, since the reference is indirect, and address it later.  Your call.

Patch-set: 1