Update patch set 2
Patch Set 2:

This opens the question as to whether we keep this MSB code or not. It is not a very active project and in light of K8S seems pretty useless. Nothing uses this code right now so why bother with it if its the only thing that brings in Jackson.

Patch-set: 2
Label: Code-Review=0