Update patch set 2
Patch Set 2:

(1 comment)

> Probably there should also be counterpart changes in the brmsgw
 > source code as they have dependency.json when they put together the
 > artifact, probably that should be changed as well ..

The exclusion lists in the policy/engine poms should probably be kept in sync.  However, it is not necessary to make any changes to them as a result of this review; the policy/engines poms are independent and policy/engine still builds fine with these changes.

Patch-set: 2