Removed null custom coders
Some code still referenced the customJacksonCoder() method, which
has been removed from drools-pdp.  However, all the code did was
set the jackson & gson coders to null, which is unnecessary, as
that is the default value for both.  Thus the code was removed
altogether so-as not to cause issues in the future.

Change-Id: Ie6f6bc8360890485ca50a751477650af7387b397
Issue-ID: POLICY-1429
Signed-off-by: Jim Hahn <jrh3@att.com>