Update patch set 3
Patch Set 3:

> Jorge/Daniel - I'm concerned that this makes these model-impl very
 > specific to Drools controller. We are about to move these
 > model-impl into the policy/models repo so that they are accessible
 > to all 3 PDP engines. Does these logging changes make that
 > difficult?

That's a good point Pam. I guess it depends on what the team is planning on doing in terms of logging conventions. The NetLoggerUtil allows 2 use cases, log() which abstracts the "LoggerFactory.getLogger()" call on the network logger and invokes before/after hooks of features for logging if any exist or getNetworkLogger() which returns the network logger itself. log() follows a drools specific kind of pattern of logging messages for the network log (Ex: [IN|DMAAP|TOPIC] <json-message>) but getNetworkLogger() allows you to dictate what the message will be since it is just a standard slf4j logger. 

My question would be is network log going to be used for all PDPs? If so then I think getNetworkLogger() is what we want to be calling here if policy/models is PDP agnostic and logging message formats may be different between PDPs. If the network log is going away then I think these reviews I have out should be put on hold to better fit the changes that are forthcoming.

Patch-set: 3
Reviewer: Gerrit User 1677 <1677@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0