Update patch set 3
Patch Set 3:

That's a good point, Pam.   I think it should be OK.  Here's my thought process on this:

One decision that it relates to the porting is what it is to be ported, if it is just the models with no networking, and may be that's how it should be, the *Manager.java classes that are touched here should be moved to somewhere else, maybe the "rest" module that is already there.

Regardless if we are moving the "rest" module within model-impl or not, the networking classes, ie. *Manager.java probably don't belong there, as they are not a model implementation, and should be moved.   That's something will have to be decided when porting.   In this regard, the "appc", "events" classes they are good as the model is kept separate from the implementation as they are placed in the dmaap bus.

The "network" logger is invoked right at the point the REST requests and only then, so if the "policy component" is only interested in the POJOS and not the networking functionality should be no impact.  

Another consideration is that the current components xacml-pdp, policy-api, policy-pap .. are using policy/commons: policy-endpoints for their dmaap needs, so in terms of dependencies there is nothing new added.

One more point, "network" logger is heavily used by developers and testers.   In integration calls, they just monitor this file to make sense of the flows.   It would be good to keep it consistent across components with same format, etc ..   We could even collect network logs across components and integrate them together if we need to understand what's going on in our policy subsystem, and they will have the same format.   Note that policy-api, policy-pap, drools-pdp, etc .. since they are using policy/common:policy-endpoints will have a network log with the same format printing the dmaap messages (done within policy-endpoints).

Patch-set: 3