Update patch set 3
Patch Set 3: Code-Review+1

Jorge, I understand people like to look at the logs. They *could* be using the Dmaap notification topic. That's what its for. What we need is for people to accurately return back error information from either the dmaap messages or REST API's.

Liam, if you are ok we can bring this in and determine if there's needs to be a strategy for different PDPs regarding logging in the common models.

Per the first issue, I was thinking about this and maybe we clean up the API implementations and build an interface that they must inherit  from that supports the notification messages (with appropriate error message details for debugging) for Policy Notification Topic. That way we can set a requirement that way which would re-usable for each PDP engine.

Patch-set: 3
Label: Code-Review=+1