Update patch set 3
Patch Set 3:

I agree with you Pam, the DMaaP notifications over POLICY-CL-MGT is the main source of analyzing what's going on in the system and provides a global view of the system.   It is very important to get those right and reporting accurately.  This is a complementary and secondary aid that I think still provides value as the granularity is on a per node basis, mainly for debugging.

Patch-set: 3