Update patch set 8
Patch Set 8:

Pramod, I think the issue of JAXBContent is not fully resolved.  We want to avoid the call to createContext on a per method basis, instead we want it on a per class basis, a global if you will, something sort of, or any other way that allows you to keep global references.

so the call to

JAXBContextFactory.createContext(new Class[] {Vserver.class, GenericVnf.class, VfModule.class,
            CloudRegion.class, ServiceInstance.class, Tenant.class}, properties); 

should happen as minimally as possible, one way is to keep this JAXBContext as class level data via static declaration.

Patch-set: 8
Label: Code-Review=0