Update patch set 2
Patch Set 2:

> One other thing, I believe we can delete the template.demo.clc now
 > that we will be able to support CLC fully in the XACML engine.
 > Possibly need to merge those changes into template.demo. Let's
 > discuss that today.

I'm fine with that and happy to discuss whenever. I've got one meeting on my calendar that Vijay doesn't want cancelled, aside from that I should be available. One note about template.demo. When doing the cleanup, it appeared that while the frequency guard is running in the demo, the only branch tested is Permit and the code can pass all tests even if the PIP bugs out. I'm not 100% on this as I didn't have time to dig in, but perhaps something to look into for later.

Patch-set: 2