New BUILD_VERSION needed as fixed in parent image
BUILD_VERSION could not be used in ENV variable
as it is cooked in the base image itself.  Need to
create a new one to reflect the drools build to
get by.

Some minor additional changes as to exit shell when
a docker command fails for clarity when executing
these operations is added as well.

Change-Id: Ic8dfc03bc0563c088d55b1991a0f8c431ab0fab2
Issue-ID: INT-998
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>