Update patch set 6
Patch Set 6:

following on my previous comment, better to disable filtering all together in pom, since there are other  ${.} that the maven plugin may expand in error before doing the docker build (#1).

For #2:

What version of docker are you using?  different versions may have a tad different behavior.   The concern is that it may fix your environment issue but break the official.  I am not familiar with the internal details on how this fabric8 plugin works, but I would assume it invokes the "docker build" underneath, so let's make sure of compatibility. 

For #3:

Right, let's switch it, since at least there would be the changes in this release related to your drools-pdp changes to use the fabric8 maven plugin.

Patch-set: 6