Update patch set 6
Patch Set 6:

> > following on my previous comment, better to disable filtering all
 > > together in pom, since there are other  ${.} that the maven
 > plugin
 > > may expand in error before doing the docker build (#1).
 > >
 > > For #2:
 > >
 > > What version of docker are you using?  different versions may
 > have
 > > a tad different behavior.   The concern is that it may fix your
 > > environment issue but break the official.  I am not familiar with
 > > the internal details on how this fabric8 plugin works, but I
 > would
 > > assume it invokes the "docker build" underneath, so let's make
 > sure
 > > of compatibility.
 > >
 > > For #3:
 > >
 > > Right, let's switch it, since at least there would be the changes
 > > in this release related to your drools-pdp changes to use the
 > > fabric8 maven plugin.
 > 
 > #2: The docker version I am using is Version 2.0.0.3.
 > #3: Sure will do that. I will extract an argument for the base
 > image version and supply that through the pom file
 > 
 > Regarding the image tagging, I can retain both min-max and
 > min-max-patch versions, but will wait for Pam or you to confirm

I see, not sure how that version compares with 

ubuntu@onap-orch-3:~$ docker --version
Docker version 18.09.6, build 481bc77

I think though, one difference with the previous way to build it was that the "src/main/docker/docker-entrypoint.sh" ended up after the same location under target but with executable permission .. and then the docker build will take place with the right permission.   Is it possible modifying the poms to do the equivalent as it was before?   If possible, it would be preferable ..   I see in the doc for the plugin some flags with permissions but I am not that familiar with the plugin.

Yeah, let's wait for Pam to see the reqs for the tagging also keeping besides your changes the x.y.z-label format.

Patch-set: 6
Label: Code-Review=0