Update patch set 3
Patch Set 3:

BTW - Jorge I think that what we will do for Frankfurt is to require the TOSCA operational policy to have constraints and requirements for these values. Hopefully we will catch it then at the API layer when we add functionality that ensures the policy created from a policy type is correct.

Patch-set: 3