Update patch set 3
Patch Set 3:

> BTW - Jorge I think that what we will do for Frankfurt is to
 > require the TOSCA operational policy to have constraints and
 > requirements for these values. Hopefully we will catch it then at
 > the API layer when we add functionality that ensures the policy
 > created from a policy type is correct.

That makes sense.

BTW: With this review, if it's left out of the yaml, or set to zero in the yaml, then the default value from the rules (currently 1500s) will be used.  Thus the only way it can be set to zero is if the rules, themselves, specify null or zero (which is currently not the case in the rules ONAP provides) AND the yaml does the same.

Patch-set: 3