Update patch set 1
Patch Set 1:

I think we could remove the <parent> section all together for this drl used in tests as well as the <dependencies> section.   I don't remember the history of the changes why it came about, but I don't think they are necessary at least at this point.   I did a clone and tried to give it a quick run with those removed but getting into some compile errors, will need to wait for Jim to synchronize the changes in the drools-applications repo first.   So, John, will need to rebase once Jim is done with the companion changes in this repo.    Once done, the test would be to remove the <parent> and <dependencies> sections from the pom.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0