Update patch set 1
Patch Set 1:

> > Its going to be very difficult to "remember" to set this value
 > > everytime we bump the version of Drools. How can we do this
 > through
 > > an environment variable for example?
 > 
 > I recall having tried to extract the tag into an ARG and supply its
 > value through the docker-maven-plugin but was not successful in
 > getting the image build going. Maybe we can create a task/story to
 > track this and I can bring along someone from my team to work on
 > this.

Not sure if there is an easy solution, even if it is handled by an ARG, it would still need to make sure that the ARG value is changed somewhere (easier may be to rely on the maven-resource-plugin filtering functionality rather than the ARG).  In case, problem and solution stays in essence the same but just moved to a different location, point being I don't see a "simple" way to infer the right version autonomously.

The reason these changes was needed are because of the checkstyle bulk changes being non-backwards compatible, otherwise could go through multiple versions of drools-apps without needing to change the drools version in the image.  This is probably a one time thing.

Same issue would be as you go up the hierarchy of images, for example, going from drools to common image, and then to base image.   At the end, you just need to know your software dependencies when you make code changes regardless is libraries or images.

I think it would be best CSITs to catch all these errors soon after merging.   What we have is minimum at this point in this area.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0