Update patch set 1
Patch Set 1: Code-Review+1

The build was successful, it just didn't show up in gerrit, so I restarted it.  In the meantime, we can go ahead and start reviewing this.
Note: until this is merged, drools-applications will not build due to the changes in the lock API.

Patch-set: 1
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1