Update patch set 1
Patch Set 1:

I see the repo is in the process of updating the Lock API which I assume has a higher priority than these changes. I'll update this review to address conflicts when the Lock updates are merged. Note that these changes won't build until until the policy models changes are merged.

Patch-set: 1