Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > (1 comment)
> 
> Thank you!

Jorge, another point -- Custom Query is not a simple GET query, meaning the first GET request won't give you back a custom query response. It is intact a two stage query, the first GET request only gives you back a URL, which is then used and modified to create the next PUT request which eventually returns a CqResponse. All this logic is handled in the AaiManager. So a direct Restful GET call to the AaiSimulator will only get you back a URL, and the StandardCoder.decode function may not work correctly. That is why I wanted to understand in a bit more detail what is this test really testing?

Patch-set: 2