Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-2
> > > 
> > > This was not what I expected at all.
> > 
> > Can you be more specific on this?
> 
> Just the guard and operations history itself is out-of-date and not compatible. What is the LCM code doing here? Looks like its specific to appc only. Now I've got 2 separate implementations of drools rules. One that sits in an "m2" submodule and the rest elsewhere. How does anyone sort this out?

Hi Pam.  The plan was to contribute the m2 code and then move the ONAP use cases to using it once it's in ONAP.  This avoids a flash cut over.  Once done, any old code can be removed from ONAP.  Regarding the other questions we will setup a meeting to review with you.

Patch-set: 1
Label: Verified=0