Update patch set 1
Patch Set 1: Code-Review-1

(27 comments)

I think in general there some areas that do not fit well with existing software.   I understand though that this is work in progress though and eventually the hope is that all gaps are bridged and we arrive to a design/implemetation that is agreed to be better by everybody.

I think it would be better to stage the work in multiple smaller reviews too, so there is more meaningful and more focused discussions.   A single one is a bit overwhelming.   I think smaller reviews would facilitate discussion, and we can come up with concrete plans of action to move forward on this work and for full integration.   Right now with this large piece of work is difficult to put the arms around it, and come up with concrete plans of action moving forward to bridge the gaps.

As an aside, one general comment is the location where this work should go, we had the controlloop module to do controlloop stuff, seems that the code should go there.   Along the same lines, it would be desirable to start with features that could be off in our system, specially for when we hit the labs, and it will also reflect that this area is in flux as some areas are very divergent for example guard, pdp-x communication with existing code.   Eventually if solid, they could be promoted to main base after being blessed by testing.

Patch-set: 1
Label: Code-Review=-1